dev-requirements.txt
dbt-duckdb

